maint: add jobs to test wheels in CI
Add jobs to the wheel build GitHub Actions workflow that will install
and test each built wheel in a fresh test runner. For each combination
of Python version and OS (currently 3.7-3.10 and OSX/Linux) this will
run a job that installs the wheel and runs the test script using the
installed wheel.

While this can give some basic confidence that the wheels do work it is
still possible that these tests can pass in CI with wheels that would
fail when transfered to another computer. At the time of this commit
that is the case with the manylinux wheels which pass all tests but fail
when installed to another Linux machine.