Work around cython parameter passing bug.
Made `__goodness` work as a cdef function.

It appears that cython has a bug in passing parameters with defaults to some
recursive functions.  `__goodness` is one of them.  In this case if any of the
parameters had defaults `parts` was passed a 0 (False).  Removing the defaults
fixes this.

Also added the `showgood` submodule to the list of submodules to doctest.