build: enable building with the limited API
Cython 3.1 makes it possible to build with the limited API. This adds
the necessary build configuration so that python-flint can be built with
the limited API which in principle makes it possible to have abi3 wheels
that can work with a range of CPython versions.

Although this builds the tests crash pretty quickly in a segfault so
more work would be needed to make these builds useful. For now we just
enable this as an experimental build mode. When it is working it would
be worth running some timings to see if it is slower than a normal
build.